+++ ChangeLog.md
@@ -2,6 +2,16 @@
 
 ## Versions ##
 
+### 0.10.0 ###
+
+* __BREAKING CHANGE__ - XML output to be deprecated in a future release (#102)
+* Adds inline paragraph styling runs - see Example38.ps1 (#75, #97)
+  * __BREAKING CHANGE__ - XML paragraph output formatting changed
+* Adds `about_PScriboExamples` help file (#103)
+  * Run `Get-Help about_PScriboExamples` to return example documentation
+* Adds dynamic `-Format` parameter to `Export-Document`
+* Adds `about_TextPlugin`, `about_HtmlPlugin` and `about_WordPlugin` help topics
+
 ### 0.9.1 ###
 
 * Fixes Word table width bug with landscape orientation